+++ apps/lib/cmp_mock_srv.c
@@ -10,7 +10,7 @@
 
 #include "apps.h"
 #include "cmp_mock_srv.h"
-#include "../../crypto/cmp/cmp_local.h" /* for access to msg->protection */
+#include "../../cmp/cmp_local.h" /* for access to msg->protection */
 #include "crypto/asn1.h"
 
 #include <openssl/cmp.h>
