+++ build.info
@@ -32,6 +32,9 @@
 -}
 
 SUBDIRS=crypto ssl apps util fuzz providers doc
+IF[{- !$disabled{cmp} -}]
+  SUBDIRS=cmp
+ENDIF
 IF[{- !$disabled{tests} -}]
   SUBDIRS=test
 ENDIF
