+++ cmp/cmp_asn.c
