+++ cmp/cmp_client.c
