+++ cmp/cmp_ctx.c
