+++ cmp/cmp_genm.c
