+++ cmp/cmp_hdr.c
