+++ cmp/cmp_local.h
