+++ cmp/cmp_msg.c
