+++ cmp/cmp_protect.c
