+++ cmp/cmp_server.c
