+++ cmp/cmp_status.c
