+++ cmp/cmp_util.c
