+++ cmp/cmp_vfy.c
