+++ crypto/build.info
@@ -5,7 +5,7 @@ SUBDIRS=objects buffer bio stack lhash hashtable rand evp asn1 pem x509 conf \
         md2 md4 md5 sha mdc2 ml_kem hmac ripemd whrlpool poly1305 \
         siphash sm3 des aes rc2 rc4 rc5 idea aria bf cast camellia \
         seed sm4 chacha modes bn ec rsa dsa dh sm2 dso \
-        err comp http ocsp cms ts srp cmac ct async ess crmf cmp encode_decode \
+        err comp http ocsp cms ts srp cmac ct async ess encode_decode \
         ffc hpke thread lms ml_dsa slh_dsa rbtree
 
 LIBS=../libcrypto
