+++ util/mkerr.pl
@@ -132,7 +132,7 @@ if ( defined $emit ) {
 } elsif ( $internal ) {
     die "Cannot mix -internal and -static\n" if $static;
     die "Extra parameters given.\n" if @ARGV;
-    @source = ( glob('crypto/*.c'), glob('crypto/*/*.c'),
+    @source = ( glob('crypto/*.c'), glob('crypto/*/*.c'), glob('cmp/*.c'),
                 glob('ssl/*.c'), glob('ssl/*/*.c'), glob('ssl/*/*/*.c'),
                 glob('providers/*.c'), glob('providers/*/*.c'),
                 glob('providers/*/*/*.c') );
